nolleyal's patch to create the MySQL test database if it does not exist.
This changes the behavior so that the test tries to create the database
rather than skipping the tests if the database does not already exist.
The database is dropped after the tests are run.

The test database name was also changed from 'test' to '_pygrtestdb' on
the grounds that the latter is less likely to conflict with existing
databases.