cleanup and refactoring of cache code.
The patch to NLMSASlice that fixes the memory leak also eliminates the
need to retain deallocID and thus eliminates cacheProxyDict as well.
The resulting code cleanup pointed out that the cache owner object,
DummyProxy, was never actually re-used, so cacheHint was refactored to
create it by default and no longer requires it to be passed in.
cacheHint now *returns* it by default, in case we want to reuse it later
on to store more cache hints under the same owner.

Renamed DummyProxy to _CacheDictOwnerObject and moved it above the only
class that now uses it.

Tests refactored accordingly.